full_examples/fraud_transactions_with_llm/requirements.txt
chalkpy[runtime,postgresql]
google-generativeai
